Create README.md
With the constant increase in vehicular traffic, existing traffic management solutions have become inefficient. Urbanization has led to an increase in traffic jams and accidents in major cities. In order to accommodate the growing needs of transport systems today, there is a need for an Intelligent Transport System. Vehicular Adhoc Network (VANET) is a growing technology that assists in Intelligent Transport Systems. VANETs enable communication between vehicles as well as fixed infrastructure called Road Side Units (RSU). We propose a distributed, collaborative traffic congestion detection and dissemination system that uses VANET. Each of the driver's smart phones is equipped with a Traffic App which is capable of location detection through Geographic Position based System (GPS). This information is relayed to a remote server which detects traffic congestion. Once congestion is confirmed the congestion information is disseminated to the end user phone through RSUs. The Mobile App transmits the location information at periodic intervals. Using the latitude, longitude and the current time, the location of each vehicle is traced. Using location information, the distance moved by the vehicle at a given time is monitored. If the value is below a fixed threshold, congestion is suspected in a particular area. If many vehicles in the same area send similar messages, traffic congestion is confirmed. Once traffic congestion is confirmed, the vehicles approaching the congested area are informed about the traffic through display boards that are available in the nearest RSUs (traffic signals). The congestion information is also made available through the Mobile App present in vehicles approaching the congested area. The approaching vehicles may take diversion and alleviate congestion. Thus a smart traffic congestion detection and dissemination system is developed to divert incoming vehicles and reduce congestion without human intervention. This system can be further enhanced by communicating the congestion information directly between vehicles